recommendation/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bleach==6.1.0
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.0
frozenlist==1.4.0
h11==0.14.0
idna==2.8
joblib==1.3.2
multidict==6.0.4
numpy==1.24.4
pandas==2.0.3
pydantic==2.4.2
pydantic-core==2.10.1
pyreqs==0.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.22.0
scikit-learn==1.3.1
scikit-surprise==1.1.3
scipy==1.10.1
sh==2.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
surprise==0.1
threadpoolctl==3.2.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.25.11
uvicorn==0.23.2
webencodings==0.5.1
yarl==1.9.2
